+++ LearnSpace/Areas/Teacher/Controllers/AssignmentController.cs
@@ -16,6 +16,10 @@ namespace LearnSpace.Web.Areas.Teacher.Controllers
         [HttpGet]
         public async Task<IActionResult> AllAssignmentsClassTeacher(int classId)
         {
+            if (!(await assignmentService.ClassExistsByIdAsync(classId)))
+            {
+                return RedirectToAction("Error404", "Error");
+            }
             var assignments = await assignmentService.GetAllAssignmentsByClassForTeacherAsync(GetUserId(), classId);
 
             return View(assignments);
@@ -24,14 +28,22 @@ namespace LearnSpace.Web.Areas.Teacher.Controllers
         [HttpGet]
         public async Task<IActionResult> AssignmentInfoForTeacher(int id)
         {
+            if (!(await assignmentService.ExistsByIdAsync(id)))
+            {
+                return RedirectToAction("Error404", "Error");
+            }
             var assignment = await assignmentService.GetAssignmentInfoForTeacherAsync(id);
 
             return View(assignment);
         }
 
         [HttpGet]
-        public IActionResult CreateAssignment(int classId)
+        public async Task<IActionResult> CreateAssignment(int classId)
         {
+            if (!(await assignmentService.ClassExistsByIdAsync(classId)))
+            {
+                return RedirectToAction("Error404", "Error");
+            }
             var model = new CreateAssignmentFormModel
             {
                 ClassId = classId
@@ -42,6 +54,10 @@ namespace LearnSpace.Web.Areas.Teacher.Controllers
         [HttpPost]
         public async Task<IActionResult> CreateAssignment(CreateAssignmentFormModel model)
         {
+            if (!ModelState.IsValid)
+            {
+                return View(model);
+            }
             await assignmentService.CreateAssignment(model);
 
             return RedirectToAction(nameof(AllAssignmentsClassTeacher), new { classId = model.ClassId });
@@ -50,9 +66,13 @@ namespace LearnSpace.Web.Areas.Teacher.Controllers
         [HttpPost]
         public async Task<IActionResult> DeleteAssignment(int id)
         {
-            var classId =await assignmentService.DeleteAssignment(id);
+            if (!(await assignmentService.ExistsByIdAsync(id)))
+            {
+                return RedirectToAction("Error404", "Error");
+            }
+            var classId = await assignmentService.DeleteAssignment(id);
 
-            return RedirectToAction(nameof(AllAssignmentsClassTeacher), new { classId =  classId});
+            return RedirectToAction(nameof(AllAssignmentsClassTeacher), new { classId = classId });
         }
     }
 }
